+++ packages/terra-section-header/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Changed
+  * Renamed title prop to text to avoid conflict with native html attribute
+
+* Added
+  * Added button role, aria-label and aria-expanded attribute to `Arrange` component wrapper
+
 ## 2.56.0 - (July 14, 2022)
 
 * Fixed